Rename template engine project to final name
