+++ dotnet/src/Extensions/TemplateEngine.PromptTemplateEngine/Blocks/ValBlock.cs
@@ -3,7 +3,7 @@
 using Microsoft.Extensions.Logging;
 using Microsoft.SemanticKernel.Orchestration;
 
-namespace Microsoft.SemanticKernel.TemplateEngine.Blocks;
+namespace Microsoft.SemanticKernel.TemplateEngine.Prompt.Blocks;
 
 internal sealed class ValBlock : Block, ITextRendering
 {
