+++ dotnet/src/Extensions/TemplateEngine.PromptTemplateEngine/PromptTemplateEngine.cs
@@ -9,9 +9,12 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.SemanticKernel.Diagnostics;
 using Microsoft.SemanticKernel.Orchestration;
-using Microsoft.SemanticKernel.TemplateEngine.Blocks;
+using Microsoft.SemanticKernel.TemplateEngine.Prompt;
+using Microsoft.SemanticKernel.TemplateEngine.Prompt.Blocks;
 
+#pragma warning disable IDE0130 // Namespace does not match folder structure
 namespace Microsoft.SemanticKernel.TemplateEngine;
+#pragma warning restore IDE0130 // Namespace does not match folder structure
 
 /// <summary>
 /// Given a prompt, that might contain references to variables and functions:
